# lib/api-zod/src/generated/types/lifeStoryEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WanisAI API - consent-first cognitive health companion
 * OpenAPI spec version: 0.1.0
 */
import type { LifeStoryEntrySource } from './lifeStoryEntrySource';

export interface LifeStoryEntry {
  id: number;
  source: LifeStoryEntrySource;
  content: string;
  createdAt: Date;
}